TrackMessages and Story00A update
fixed #431
(but that was not the main purpose)

On example of packet 5, a new and universal structure for this library
has been set up.

- Trackside input now structurally SRS conformal (main difference is
inclusion of NID_PACKET and the structure of the N_ITER sections.

Packet 5 is now extracted in full, and a separate operator provided for
Calculate Train Position.