Story00A and Amsterdam- Utrecht fixed dependencies
Story00A is already relying on “real” balises from AmsterdamUtrecht
(section Bijlmer).
As we need real data for AmsterdamUtrecht, I had to set the Q_DIR for
Packet 5 in BG364 to the correct value.
But in Story00A we need “both directions” so the message can pass the
balise channel. This has now been done.